Automating Vuex Modules
