+++ src/frontend/src/modules/index.js
@@ -1,6 +1,10 @@
-import Builder from "./builder/Builder";
-import Cart from "./cart/Cart";
-import Profile from "./profile/Profile";
-import Orders from "./orders/Orders";
+const requireContext = require.context("@/modules/", true, /store\.js$/);
 
-export { Builder, Cart, Profile, Orders };
+export default requireContext.keys().reduce((modules, filename) => {
+  const moduleName = filename
+    .split("/")[1]
+    .replace(/^\w/, (c) => c.toUpperCase());
+  modules[moduleName] =
+    requireContext(filename).default || requireContext(filename);
+  return modules;
+}, {});
