--- src/frontend/src/store/modules/index.js
@@ -1,11 +1,0 @@
-import Auth from "@/store/modules/auth.store.js";
-import Builder from "@/store/modules/builder.store.js";
-import Cart from "@/store/modules/cart.store.js";
-import Orders from "@/store/modules/orders.store.js";
-
-export default {
-  Auth,
-  Builder,
-  Cart,
-  Orders,
-};
