+++ src/frontend/src/views/Cart/__tests__/Cart.spec.js
@@ -2,7 +2,7 @@ import { mount, createLocalVue } from "@vue/test-utils";
 import Vuex from "vuex";
 import { generateMockStore } from "@/store/mocks";
 import { SET_ENTITY } from "@/store/mutation-types";
-import Cart from "../Cart";
+import Cart from "../Index";
 import {
   CartFooterView,
   PizzasListLayout,
