+++ src/frontend/src/views/Cart/components/CartFooterView.vue
