+++ src/frontend/src/views/Cart/components/PizzasListLayout.vue
