+++ src/frontend/src/views/Cart/components/__tests__/CartFooterView.spec.js
