+++ src/frontend/src/views/Index/components/BuilderDoughSelector.vue
