+++ src/frontend/src/views/Index/components/BuilderIngredientsSelector.vue
