+++ src/frontend/src/views/Index/components/BuilderPizzaView.vue
