+++ src/frontend/src/views/Index/components/BuilderSauceSelector.vue
