+++ src/frontend/src/views/Index/components/BuilderSizeSelector.vue
