+++ src/frontend/src/views/Index/components/__tests__/BuilderIngredientsSelector.spec.js
