+++ src/frontend/src/views/Index/components/__tests__/BuilderPizzaView.spec.js
