+++ src/frontend/src/views/Index/components/__tests__/BuilderSauceSelector.spec.js
