--- src/frontend/src/views/__tests__/Cart.spec.js
@@ -1,9 +1,0 @@
-import { shallowMount } from "@vue/test-utils";
-import Cart from "../Cart";
-
-describe("Cart", () => {
-  it("is rendered", () => {
-    const wrapper = shallowMount(Cart);
-    expect(wrapper.exists()).toBeTruthy();
-  });
-});
