--- src/frontend/src/views/__tests__/Main.spec.js
@@ -1,11 +1,0 @@
-import { shallowMount } from "@vue/test-utils";
-import Main from "../Index";
-
-const stubs = ["router-view"];
-
-describe("Main", () => {
-  it("is rendered", () => {
-    const wrapper = shallowMount(Main, { stubs });
-    expect(wrapper.exists()).toBeTruthy();
-  });
-});
