--- src/frontend/src/views/__tests__/Orders.spec.js
@@ -1,9 +1,0 @@
-import { shallowMount } from "@vue/test-utils";
-import Orders from "../Orders";
-
-describe("Orders", () => {
-  it("is rendered", () => {
-    const wrapper = shallowMount(Orders);
-    expect(wrapper.exists()).toBeTruthy();
-  });
-});
