--- src/frontend/src/views/__tests__/Profile.spec.js
@@ -1,9 +1,0 @@
-import { shallowMount } from "@vue/test-utils";
-import Profile from "../Profile";
-
-describe("Profile", () => {
-  it("is rendered", () => {
-    const wrapper = shallowMount(Profile);
-    expect(wrapper.exists()).toBeTruthy();
-  });
-});
